+++ pygazpar/__init__.py
@@ -1,1 +1,2 @@
 from pygazpar.client import Client
+from pygazpar.client import LoginError